Analysis of data taken from NASA to analyze data on all meteorites falling on Earth
Analysis of the data set taken from NASA’s international agency for the history of meteorites falling on Earth to analyze this data and extract a huge amount of information from it. This data is suitable for entering it into machine learning models to make decisions and make predictions based on them.